backend/requirements.txt
django>=4.2
djangorestframework
psycopg2-binary
django-cors-headers
gunicorn
djangorestframework-simplejwt
